Update outdated tests
