+++ tests/unit/experiment_api/test_request_exception_wrapper.py
@@ -9,13 +9,13 @@ from comet_llm.exceptions import exceptions
 from comet_llm.experiment_api import request_exception_wrapper
 
 
-@pytest.fixture(autouse=True)
+@pytest.fixture()
 def mock_imports(patch_module):
     patch_module(request_exception_wrapper, "config")
     patch_module(request_exception_wrapper, "failed_response_handler")
 
 
-def test_wrap_no_exceptions():
+def test_wrap_no_exceptions(mock_imports):
     @request_exception_wrapper.wrap()
     def f():
         return "return-value"
@@ -23,7 +23,7 @@ def test_wrap_no_exceptions():
     assert f() == "return-value"
 
 
-def test_wrap__request_exception_caught__comet_exception_raised():
+def test_wrap__request_exception_caught__comet_exception_raised(mock_imports):
     @request_exception_wrapper.wrap()
     def f():
         raise requests.RequestException
@@ -32,7 +32,7 @@ def test_wrap__request_exception_caught__comet_exception_raised():
         f()
 
 
-def test_wrap__on_prem_check_enabled__request_exception_caught__on_prem_detected__comet_exception_raised_with_additional_message():
+def test_wrap__on_prem_check_enabled__request_exception_caught__on_prem_detected__comet_exception_raised_with_additional_message(mock_imports):
     @request_exception_wrapper.wrap(check_on_prem=True)
     def f():
         raise requests.RequestException
@@ -43,7 +43,7 @@ def test_wrap__on_prem_check_enabled__request_exception_caught__on_prem_detected
             f()
 
 
-def test_wrap__on_prem_check_enabled__request_exception_caught__on_prem_not_detected__comet_exception_raised_without_additional_message():
+def test_wrap__on_prem_check_enabled__request_exception_caught__on_prem_not_detected__comet_exception_raised_without_additional_message(mock_imports):
     @request_exception_wrapper.wrap(check_on_prem=True)
     def f():
         raise requests.RequestException
@@ -56,11 +56,18 @@ def test_wrap__on_prem_check_enabled__request_exception_caught__on_prem_not_dete
             f()
 
 
-def test_wrap__request_exception_with_not_None_response__exception_handled_by_failed_response_handler():
-    exception = requests.RequestException(response=box.Box(text="some-text"))
+def test_wrap__request_exception_non_llm_project_sdk_code__log_specifc_message_in_exception():
+    exception = requests.RequestException()
+    exception.response = box.Box(text=json.dumps({"sdk_error_code": 34323}))
 
-    @request_exception_wrapper.wrap()
+    expected_log_message = "Failed to send prompt to the specified project as it is not an LLM project, please specify a different project name."
+
+    @request_exception_wrapper.wrap(check_on_prem=True)
     def f():
         raise exception
 
+
+    with pytest.raises(exceptions.CometLLMException) as excinfo:
         f()
+
+    assert excinfo.value.args == (expected_log_message, )
